Extract 'Selected soldier' from adapter to fragment
