--- BF4Intel/src/main/res/layout/side_soldier_box.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Spinner
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  style="@android:style/Widget.Holo.Spinner"
-  android:id="@+id/spinner_soldier"
-  android:layout_width="match_parent"
-  android:layout_height="wrap_content"
-  android:spinnerMode="dropdown"
-  />
