+++ BF4Intel/src/main/res/values/styles.xml
@@ -55,6 +55,7 @@
   <style name="TextAppearance.Holo.MediumLarge" parent="@android:style/TextAppearance.Holo.Medium">
     <item name="android:textSize">@dimen/text_ml</item>
   </style>
+
   <style name="TextAppearance.Holo.XLarge" parent="@android:style/TextAppearance.Holo.Large">
     <item name="android:textSize">@dimen/text_xl</item>
   </style>
@@ -273,6 +274,18 @@
     <item name="android:clipToPadding">false</item>
     <item name="android:padding">@dimen/spacing_m</item>
     <item name="android:scrollbarStyle">outsideOverlay</item>
+  </style>
+
+  <style name="NavigationDrawerHeading">
+    <item name="android:paddingTop">@dimen/spacing_m</item>
+    <item name="android:paddingRight">@dimen/spacing_m</item>
+    <item name="android:paddingBottom">@dimen/spacing_s</item>
+    <item name="android:paddingLeft">@dimen/spacing_m</item>
+    <item name="android:layout_width">match_parent</item>
+    <item name="android:layout_height">wrap_content</item>
+    <item name="android:textAppearance">@android:style/TextAppearance.Holo.Small.Inverse</item>
+    <item name="android:textAllCaps">true</item>
+    <item name="android:textColor">@color/lightgrey</item>
   </style>
 
 </resources>
